guardrail/pyproject.toml
[project]
name = "guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # ---- Core app ----
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",

    # ---- ML / LLM ----
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "huggingface-hub>=0.36.0",
    "tiktoken==0.12.0",
    "regex==2026.1.15",
    "jiter==0.12.0",

    # ---- Validation / typing ----
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "annotated-types==0.7.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "types-requests==2.32.4.20260107",

    # ---- HTTP stack ----
    "httpx==0.28.1",
    "httpcore==1.0.9",
    "anyio==4.12.1",
    "h11==0.16.0",
    "sniffio==1.3.1",
    "urllib3==2.6.3",
    "charset-normalizer==3.4.4",
    "idna==3.11",
    "certifi==2026.1.4",

    # ---- Azure ----
    "azure-core==1.38.0",
    "azure-common==1.1.28",
    "azure-search-documents==11.6.0",
    "isodate==0.7.2",

    # ---- Utils ----
    "tqdm==4.67.1",
    "cached-property==2.0.1",
    "PyPDF2==3.0.1",
    "distro==1.9.0",

    # ---- Portkey ----
    "portkey-ai==2.1.0",
]